% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bmt_model}
\alias{bmt_model}
\title{Create a \code{bmt_model}}
\usage{
bmt_model(
  discharge_model,
  admission_model,
  admission_surv_data = default_admission_surv_data()
)
}
\arguments{
\item{discharge_model}{discharge model}

\item{admission_model}{admission model}

\item{admission_surv_data}{survival curves for simulating admission-discharge process}
}
\description{
Create a \code{bmt_model}
}
