% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{control_sim}
\alias{control_sim}
\title{Admission-discharge simulation control}
\usage{
control_sim(n_iter = 100, h_trunc = 300)
}
\arguments{
\item{n_iter}{number of iterations to use in simulation}

\item{h_trunc}{maximum steps ahead for simulating discharges}
}
\description{
Admission-discharge simulation control
}
\details{
\code{h_trunc} should be large enough that survival probabilities are approximately zero.
Setting it too low will underestimate census.
}
