% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_transplant_types}
\alias{default_transplant_types}
\title{Default transplant types}
\usage{
default_transplant_types(add = NULL, remove = NULL)
}
\arguments{
\item{add}{vector of transplant types to add}

\item{remove}{vector of transplant types to remove}
}
\description{
Default transplant types
}
