% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extend_discharge_data}
\alias{extend_discharge_data}
\title{Convert encounter-level data to person-period format}
\usage{
extend_discharge_data(data, h, end_var = NULL)
}
\arguments{
\item{data}{dataframe minimally containing \code{time}}

\item{h}{number of periods to extend data}

\item{end_var}{variable denoting discharge time}
}
\description{
Convert encounter-level data to person-period format
}
\details{
This function assumes covariates are not time-varying and carries their values
forward
}
