% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_admission}
\alias{predict_admission}
\title{Forecast admissions}
\usage{
predict_admission(model, new_data)
}
\arguments{
\item{model}{admission workflow}

\item{new_data}{dataframe to predict, see \code{prep_admission_data()}}
}
\value{
\code{new_data} with \code{.value} column containing predicted admissions
}
\description{
Forecast admissions
}
