% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_census}
\alias{predict_census}
\title{Predict BMT census}
\usage{
predict_census(
  models,
  h,
  current_census,
  scheduled_admissions = NULL,
  control = control_sim(),
  type = "census"
)
}
\arguments{
\item{models}{named list containing \code{discharge}, a workflow for predicting hospital discharges;
\code{admission}, a workflow for predicting new admissions; and \code{admission_surv_data}, training data
to compute survival probabilities for new admissions}

\item{h}{numeric vector of future times to predict}

\item{current_census}{patient-level dataframe containing currently admitted patients}

\item{scheduled_admissions}{dataframe containing scheduled admissions}

\item{control}{a control object}

\item{type}{'census' or 'survival'}
}
\value{
dataframe of prediction for current census component, new admission component, and total census
}
\description{
Produce a prediction of the BMT census \code{h} days into the future using admission and discharge
models, the current census, and optionally upcoming scheduled admissions.
}
\details{
The \code{admission_surv_data} component of \code{models} is only relevant if \code{discharge} has predictors
other than \code{transplant_type}. Otherwise it can be populated with \code{default_admission_surv_data()}.
\code{current_census} must contain a \code{date} column indicating the date of the current census. \code{scheduled_admissions}
must contain:
\itemize{
\item \code{date}, the day for which admissions are scheduled
\item \code{transplant_type}
\item \code{scheduled}, the number of admissions, possibly 0
}
}
