% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_discharge}
\alias{predict_discharge}
\title{Forecast current census into the future}
\usage{
predict_discharge(model, new_data, h, type = "census")
}
\arguments{
\item{model}{discharge workflow}

\item{new_data}{dataframe containing current census}

\item{h}{numeric vector of future times to predict}

\item{type}{'census' or 'survival'}
}
\value{
see details
}
\description{
Forecast current census into the future
}
\details{
If \code{type} is 'census' a dataframe containing \code{.h} and \code{.pred} is returned where \code{.pred} is the
predicted number of remaining patients at each time point. If \code{type} is 'survival' a dataframe
with the same number of rows as \code{new_data} is returned where \code{.pred} is a list column containing
survival probabilities at each time.
}
