% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{sim_admission_discharge}
\alias{sim_admission_discharge}
\title{Simulate discharges from predicted admissions}
\usage{
sim_admission_discharge(
  pred_admission,
  model,
  admission_surv_data,
  h = NULL,
  type = "census",
  control = control_sim()
)
}
\arguments{
\item{pred_admission}{admission predictions from \code{predict_admission()}}

\item{model}{discharge workflow}

\item{admission_surv_data}{data used to compute admission survival curves, likely training data}

\item{h}{numeric vector of future times to predict, inferred from \code{pred_admission} by default}

\item{type}{'census' or 'survival'}

\item{control}{a control object}
}
\value{
dataframe containing \code{.h} and \code{.pred} is returned where \code{.pred} is the
predicted number of remaining patients at each time point
}
\description{
This function takes predictions made by \code{predict_admission()} and a workflow
for predicting discharges and returns a predicted number of new admissions still
admitted for each future time, \code{h}.
}
